+++ classes/Mvc/View/TwigView.php
@@ -9,7 +9,7 @@ class TwigView extends AbstractView {
   /**
    * @var string
    */
-  protected $default_template_file_extension = 'twig';
+  protected $defaultTemplateFileExtension = 'twig';
 
   /**
    * @throws ApplicationPartMissingException
@@ -22,7 +22,7 @@ class TwigView extends AbstractView {
    * @return \Twig_Environment
    * @throws ApplicationPartMissingException
    */
-  protected function get_rendering_engine() {
+  protected function getRenderingEngine() {
     if (!isset($this->rendering_engine)) {
       if (!is_dir(Configuration::get('application', 'templates_directory'))) {
         throw new ApplicationPartMissingException('Template directory "' . Configuration::get('application', 'templates_directory') . '" does not exist.');
@@ -38,47 +38,87 @@ class TwigView extends AbstractView {
   }
 
   /**
-   * Adds a filter to use in the template
-   *
-   * @param $name string Name of the filter to use in the template
-   * @param $function string Name of the function to execute for the value from the template
+   * @param string $name Name of the filter to use in the template
+   * @param mixed $function Name of the function to execute for the value from the template
+   * @param bool $htmlEscape
    */
-  public function add_output_filter($name, $function, $htmlEscape = FALSE) {
+  public function addOutputFilter($name, $function, $htmlEscape = FALSE) {
     $options = [];
     if (!$htmlEscape) {
       $options = ['is_safe' => ['all']];
     }
-    $this->get_rendering_engine()->addFilter(new \Twig_SimpleFilter($name, $function, $options));
+    $this->getRenderingEngine()->addFilter(new \Twig_SimpleFilter($name, $function, $options));
   }
 
   /**
    * @param $name
    * @return bool
    */
-  public function has_output_filter($name) {
-    return $this->get_rendering_engine()->getFilter($name) instanceof \Twig_SimpleFilter;
+  public function hasOutputFilter($name) {
+    return $this->getRenderingEngine()->getFilter($name) instanceof \Twig_SimpleFilter;
   }
 
   /**
    * Adds a function to use in the template
    *
-   * @param $name string Name of the function to use in the template
-   * @param $function string Name of the function to execute for the value from the template
+   * @param string $name string Name of the function to use in the template
+   * @param mixed $function string Name of the function to execute for the value from the template
+   * @param bool $htmlEscape
    */
-  public function add_output_function($name, $function, $htmlEscape = FALSE) {
+  public function addOutputFunction($name, $function, $htmlEscape = FALSE) {
     $options = [];
     if (!$htmlEscape) {
       $options = ['is_safe' => ['all']];
     }
-    $this->get_rendering_engine()->addFunction(new \Twig_SimpleFunction($name, $function, $options));
+    $this->getRenderingEngine()->addFunction(new \Twig_SimpleFunction($name, $function, $options));
   }
 
   /**
-   * @param $name
+   * @param string $name
    * @return bool
    */
+  public function hasOutputFunction($name) {
+    return $this->getRenderingEngine()->getFunction($name) instanceof \Twig_SimpleFunction;
+  }
+
+  /**
+   * @param string $name Name of the filter to use in the template
+   * @param mixed $function Name of the function to execute for the value from the template
+   * @param bool $htmlEscape
+   * @deprecated Since: 1.4, Removal: 1.5, Reason: Use ->addOutputFilter() instead
+   */
+  public function add_output_filter($name, $function, $htmlEscape = FALSE) {
+    $this->addOutputFilter($name, $function, $htmlEscape);
+  }
+
+  /**
+   * @param string $name
+   * @return bool
+   * @deprecated Since: 1.4, Removal: 1.5, Reason: Use ->hasOutputFilter() instead
+   */
+  public function has_output_filter($name) {
+    return $this->hasOutputFilter($name);
+  }
+
+  /**
+   * Adds a function to use in the template
+   *
+   * @param string $name string Name of the function to use in the template
+   * @param mixed $function string Name of the function to execute for the value from the template
+   * @param bool $htmlEscape
+   * @deprecated Since: 1.4, Removal: 1.5, Reason: Use ->addOutputFunction() instead
+   */
+  public function add_output_function($name, $function, $htmlEscape = FALSE) {
+    $this->addOutputFunction($name, $function, $htmlEscape);
+  }
+
+  /**
+   * @param string $name
+   * @return bool
+   * @deprecated Since: 1.4, Removal: 1.5, Reason: Use ->hasOutputFunction() instead
+   */
   public function has_output_function($name) {
-    return $this->get_rendering_engine()->getFunction($name) instanceof \Twig_SimpleFunction;
+    return $this->hasOutputFunction($name);
   }
 
 }