+++ tests/Unit/Mvc/AbstractControllerTest.php
@@ -6,9 +6,7 @@ use AppZap\PHPFramework\Mvc\Request;
 use AppZap\PHPFramework\Mvc\View\AbstractView;
 
 class TestResponse extends AbstractView {
-  public function __construct() {
-  }
-  protected function get_rendering_engine() {
+  protected function getRenderingEngine() {
   }
 }